+++ cmsplugin_blocks/forms/container.py
@@ -22,6 +22,7 @@ class ContainerForm(forms.ModelForm):
             "template",
             "features",
             "image",
+            "image_alt",
             "content",
         ]
         widgets = {