+++ cmsplugin_blocks/models/card.py
@@ -77,6 +77,16 @@ class Card(SmartFormatMixin, CMSPlugin):
     ``BLOCKS_ALLOWED_IMAGE_EXTENSIONS``.
     """
 
+    image_alt = models.CharField(
+        _("Alternative image text"),
+        blank=True,
+        max_length=125,
+        default="",
+    )
+    """
+    An optional text string for alternative image text.
+    """
+
     content = models.TextField(
         _("Content"),
         blank=True,