+++ cmsplugin_blocks/models/slider.py
@@ -151,6 +151,16 @@ class SlideItem(SmartFormatMixin, models.Model):
     ``BLOCKS_ALLOWED_IMAGE_EXTENSIONS``.
     """
 
+    image_alt = models.CharField(
+        _("Alternative image text"),
+        blank=True,
+        max_length=125,
+        default="",
+    )
+    """
+    An optional text string for alternative image text.
+    """
+
     content = models.TextField(
         _(u"Content"),
         blank=True,